Make blackboard tests run faster by not serializing to disk unless necessary.
